pyproject.toml
[tool.poetry]
name = "teehr"
version = "0.4.0"
description = "Tools for Exploratory Evaluation in Hydrologic Research"
authors = [
    "RTI International",
    "Matthew Denno <mdenno@rti.org>",
    "Katie van Werkhoven <kvanwerkhoven@rti.org>",
    "Sam Lamont <slamont@rti.org>",
]
license = "GNU v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^0.14.0"
duckdb = "^1.0.0"
pydantic = "^2.4.2"
hydrotools-metrics = "^1.3.3"
hydrotools-nwis-client = "^3.3.1"
dask = {extras = ["dataframe"], version = "^2024.11.1"}
fsspec = ">=2023.10.0"
ujson = "^5.8.0"
kerchunk = "^0.2.2"
gcsfs = ">=2023.10.0"
s3fs = ">=2023.10.0"
zarr = "^2.16.1"
rioxarray = "^0.15.0"
rasterio = "^1.3.9"
h5py = "3.12.1"
pyarrow = "^15.0.0"
httpx = "^0.25.1"
pandas = "^2.2.0"
pyspark = {extras = ["pandas-on-spark"], version = "^3.5.2"}
dataretrieval = "^1.0.9"
numba = "^0.60.0"
arch = "^7.0.0"
pandera = {extras = ["pyspark"], version = "^0.20.4"}
netcdf4 = "1.6.5"
bokeh = "^3.5.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"

[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
sphinx-design = "^0.5.0"
numpydoc = "^1.6.0"
pre-commit = "^3.6.0"
pre-commit-hooks = "^4.5.0"
flake8-docstrings = "^1.7.0"
myst-nb = "^1.0.0"
nbstripout = "^0.7.1"
pickleshare = "^0.7.5"
sphinx-autoapi = "3.0.0"
pydata-sphinx-theme = "^0.16.0"
sphinx-autobuild = "^2024.10.3"


[tool.numpydoc_validation]
checks = [
    "all",   # report on all checks, except the below
    "EX01",  # No examples section found
    "RT01",  # No Returns section found
    "SA01",  # See Also section not found
    "ES01",  # No extended summary found
    "GL01",  # Docstring text (summary) should start in the line immediately
    "SS06",  # Summary should fit in a single line
    "PR01",  # Parameters {missing_params} not documented
    "GL02"   # Closing quotes should be placed in the line after the last text in docstring"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
